[#835] split model and uimodel
