+++ common/src/commonMain/kotlin/com/oztechan/ccc/common/api/model/CurrencyResponse.kt
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) 2021 Mustafa Ozhan. All rights reserved.
  */
-package com.oztechan.ccc.common.entity
+package com.oztechan.ccc.common.api.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -9,8 +9,8 @@ import kotlinx.serialization.json.JsonNames
 
 @Suppress("OPT_IN_USAGE")
 @Serializable
-internal data class CurrencyResponseEntity(
+internal data class CurrencyResponse(
     @JsonNames("base", "base_code") var base: String,
     @SerialName("date") var date: String? = null,
-    @JsonNames("rates", "conversion_rates") var rates: RatesEntity
+    @JsonNames("rates", "conversion_rates") var rates: Rates
 )
