+++ common/src/commonMain/kotlin/com/oztechan/ccc/common/api/service/ApiServiceImpl.kt
@@ -4,7 +4,7 @@
 package com.oztechan.ccc.common.api.service
 
 import com.oztechan.ccc.common.BuildKonfig
-import com.oztechan.ccc.common.entity.CurrencyResponseEntity
+import com.oztechan.ccc.common.api.model.CurrencyResponse
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.get
@@ -14,7 +14,7 @@ import io.ktor.http.takeFrom
 
 internal class ApiServiceImpl(private val client: HttpClient) : ApiService {
 
-    override suspend fun getRatesByBackend(base: String): CurrencyResponseEntity = client.get {
+    override suspend fun getRatesByBackend(base: String): CurrencyResponse = client.get {
         url {
             takeFrom(BuildKonfig.BASE_URL_BACKEND)
             path(PATH_CURRENCY, PATH_BY_BASE)
@@ -22,7 +22,7 @@ internal class ApiServiceImpl(private val client: HttpClient) : ApiService {
         }
     }.body()
 
-    override suspend fun getRatesByAPI(base: String): CurrencyResponseEntity = client.get {
+    override suspend fun getRatesByAPI(base: String): CurrencyResponse = client.get {
         url {
             takeFrom(BuildKonfig.BASE_URL_API)
             path(PATH_LATEST)
@@ -30,7 +30,7 @@ internal class ApiServiceImpl(private val client: HttpClient) : ApiService {
         }
     }.body()
 
-    override suspend fun getRatesByPremiumAPI(base: String): CurrencyResponseEntity = client.get {
+    override suspend fun getRatesByPremiumAPI(base: String): CurrencyResponse = client.get {
         url {
             takeFrom(BuildKonfig.BASE_URL_API_PREMIUM)
             path(PATH_PREMIUM_VERSION, BuildKonfig.API_KEY_PREMIUM, PATH_LATEST, base)
