+++ common/src/commonTest/kotlin/com/oztechan/ccc/common/mapper/CurrencyResponseTest.kt
@@ -1,12 +1,12 @@
 package com.oztechan.ccc.common.mapper
 
-import com.oztechan.ccc.common.entity.CurrencyResponseEntity
-import com.oztechan.ccc.common.entity.RatesEntity
 import com.oztechan.ccc.common.model.CurrencyResponse
 import com.oztechan.ccc.common.model.Rates
 import com.oztechan.ccc.common.util.assertAllTrue
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import com.oztechan.ccc.common.api.model.CurrencyResponse as CurrencyResponseEntity
+import com.oztechan.ccc.common.api.model.Rates as RatesEntity
 
 class CurrencyResponseTest {
     @Suppress("LongMethod")
