+++ common/src/commonTest/kotlin/com/oztechan/ccc/common/mapper/RatesTest.kt
@@ -1,6 +1,6 @@
 package com.oztechan.ccc.common.mapper
 
-import com.oztechan.ccc.common.entity.RatesEntity
+import com.oztechan.ccc.common.api.model.Rates
 import com.oztechan.ccc.common.util.assertAllTrue
 import kotlin.test.Test
 
@@ -9,7 +9,7 @@ class RatesTest {
     @Suppress("LongMethod")
     @Test
     fun toModel() {
-        val entity = RatesEntity()
+        val entity = Rates()
         val model = entity.toModel()
         assertAllTrue(
             entity.base == model.base, entity.date == model.date, entity.aed == model.aed,
